[TSK-79] Re_Commit
